requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.4
django-allauth==65.4.1
django-appconf==1.0.6
django-compressor==4.5.1
django-cors-headers==4.7.0
django-storages==1.14.5
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
filelock==3.18.0
fsspec==2025.3.2
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
gunicorn==23.0.0
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
proto-plus==1.26.1
protobuf==6.30.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rcssmin==1.1.2
regex==2024.11.6
requests==2.32.3
rjsmin==1.2.2
rsa==4.9
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.8.2
